静的サーバと動的コンテンツサーバを分離
