+++ src/main/java/jp/co/topgate/teru/web/DynamicContentHandler.java
@@ -1,0 +1,17 @@
+package jp.co.topgate.teru.web;
+
+/**
+ * Created by terufumishimoji on 2016/12/16.
+ */
+public class DynamicContentHandler {
+
+    public HTTPResponse doGET(HTTPRequest request) {
+        HTTPResponse response = new HTTPResponse();
+        return response;
+    }
+
+    public HTTPResponse doPOST(HTTPRequest request) {
+        HTTPResponse response = new HTTPResponse();
+        return response;
+    }
+}
