+++ src/main/java/jp/co/topgate/teru/web/HTTPRequestHandler.java
@@ -22,63 +22,35 @@ class HTTPRequestHandler {
      */
     HTTPResponse handle(HTTPRequest request) throws Exception {
 
+        // 静的コンテンツ処理か動的コンテンツ処理かの判断をする
+        // 判断はリクエストのURIに依存する。
+        // <リクエストURIにprogramがマッチ>
+        // YES ==> 動的
+        // NO ==> 静的
+        // ブロック内でお互いのハンドラインスタンスを作ってあげれば良い。
+        // ここで問題なのが、 リクエストの状態はその後のハンドラに依存するということ
+        // つまりここでも渡してあげないといけないというめんどくささ。レスポンスの場合は
+        // ハンドラの呼び出し先のThreadでrespond処理をしているので、レスポンスの実態は
+        // ハンドラに返さないといけない。
+
         HTTPResponse response = new HTTPResponse();
 
-        if ("GET".equals(request.getRequestMethod())) {
-
-            File file = new File(request.getResourcePath());
-
-            if (file.exists() && file.isFile()) {
-                System.out.println("リソースは存在しました。");
-
-                response.setStatusCode(200);
-                response.setReasonPhrase("OK");
-                response.setHeader("Content-Type", response.getContentType(file.getName()));
-                response.setMessageBody(file);
-
+        if (request.getRequestMethod() == "program") {
+            // 動的アプリケーション
+            DynamicContentHandler dynamicContentHandler = new DynamicContentHandler();
+            // GET/POST
+            if (request.getRequestMethod() == "GET") {
+                response = dynamicContentHandler.doGET();
             } else {
-                System.out.println("リソースは存在しませんでした。");
-                response.setStatusCode(404);
-                response.setReasonPhrase("Not Found");
-                response.setHeader("Content-Type", "text/html");
-                response.setMessageBodyError(errorHandle("404"));
+                response = dynamicContentHandler.doPOST();
             }
-
         } else {
-            System.out.println("許可されていないHTTPメソッドです。");
-            response.setStatusCode(405);
-            response.setReasonPhrase("Method not allowed Explained");
-            response.setHeader("Content-Type", "text/html");
-            response.setMessageBodyError(errorHandle("405"));
+            // 静的
+            // レスポンスを受け取りたい
+            StaticContentHandler staticContentHandler = new StaticContentHandler();
+            response = staticContentHandler.doGET(request);
         }
+
         return response;
-    }
-
-    /**
-     * エラーコンテンツを文字列で返すハンドラメソッド
-     *
-     * @param errorStatus エラーステータスコード
-     * @return エラーコンテンツ
-     */
-    String errorHandle(String errorStatus) {
-        Map<String, String> errorMessage = new HashMap<String, String>() {
-            {
-                put("404", "Not Found");
-                put("405", "Method not allowed Explained");
-            }
-        };
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append("<!DOCTYPE html>\n");
-        stringBuilder.append("<html>\n");
-        stringBuilder.append("    <head>\n");
-        stringBuilder.append("        <meta charset=\"UTF-8\" />\n");
-        stringBuilder.append("        <title>Simple HTTP Server</title>\n");
-        stringBuilder.append("    <body>\n");
-        stringBuilder.append("        <h1>" + errorStatus + " " + errorMessage.get(errorStatus) + "</h1>\n");
-        stringBuilder.append("    </body>\n");
-        stringBuilder.append("</html>");
-
-        String errorContent = new String(stringBuilder);
-        return errorContent;
     }
 }
